+++ src/main/java/com/dnd/sub/domain/auth/dto/response/TokenResponse.java
@@ -2,6 +2,6 @@ package com.dnd.sub.domain.auth.dto.response;
 
 public record TokenResponse(
     String accessToken,
-    String refreshToken) {
-
+    String refreshToken
+) {
 }
