+++ src/main/java/com/dnd/sub/domain/auth/exception/TokenException.java
@@ -1,8 +1,5 @@
-package com.dnd.sub.global.exception;
+package com.dnd.sub.domain.auth.exception;
 
-import com.dnd.sub.global.enums.GlobalErrorCode;
-import com.dnd.sub.global.enums.TokenErrorCode;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
