+++ src/main/java/com/dnd/sub/global/entity/BaseEntity.java
@@ -15,9 +15,10 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 public abstract class BaseEntity {
 
   @CreatedDate
-  @Column(updatable = false)
+  @Column(name = "created_at", updatable = false)
   private LocalDateTime createdAt;
 
   @LastModifiedDate
+  @Column(name = "updated_At")
   private LocalDateTime updatedAt;
 }
