+++ src/main/java/com/dnd/sub/global/security/dto/KakaoAuthMemberDto.java
@@ -9,5 +9,4 @@ public record KakaoAuthMemberDto(
     String nickname,
     String email
 ) {
-
 }
