+++ src/main/java/com/dnd/sub/global/security/jwt/JwtFilter.java
@@ -1,11 +1,10 @@
 package com.dnd.sub.global.security.jwt;
 
 import com.dnd.sub.domain.member.repository.MemberRepository;
-import com.dnd.sub.global.enums.TokenErrorCode;
-import com.dnd.sub.global.exception.TokenException;
+import com.dnd.sub.domain.auth.exception.TokenErrorCode;
+import com.dnd.sub.domain.auth.exception.TokenException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -37,7 +36,7 @@ public class JwtFilter extends OncePerRequestFilter {
             Long memberId = jwtProvider.extractUserId(token);
 
             memberRepository.findById(memberId)
-                .orElseThrow(() -> new TokenException(TokenErrorCode.NOT_FOUND_USER));
+                .orElseThrow(() -> new TokenException(TokenErrorCode.NOT_FOUND_MEMBER));
 
             Authentication authentication = new UsernamePasswordAuthenticationToken(
                 memberId, null, Collections.emptyList());
